+++ dist-lib/Messages-CGRnj0KV.js
@@ -1,10 +1,10 @@
-import { _ as e, h as t } from "./MobileAppViewer-8sWgnhGz.js";
+import { _ as e, h as t } from "./MobileAppViewer-BTU3wt33.js";
 import { t as n } from "./utils-B6YmNDS2.js";
 import { t as r } from "./createLucideIcon--WjuKCts.js";
 import { t as i } from "./plus-Dx6tdmvY.js";
 import { t as a } from "./search-BKG97C0x.js";
 import { t as o } from "./useTranslation-BpGBgycp.js";
-import { n as s, r as c, t as l } from "./MobileSplitView-Boyggner.js";
+import { n as s, r as c, t as l } from "./MobileSplitView-BU3DwYXP.js";
 import { useEffect as u, useRef as d, useState as f } from "react";
 import { jsx as p, jsxs as m } from "react/jsx-runtime";
 var h = r("send", [["path", {